Add the boilerplate method to deserialize NewgroundsSongs
Note that generally serde-deserializers provide a crate wide from_str method that's generic over the type being deserialized. This is not possible here because that method would have no way of knowing which delimiter to use. We could of course work around that by writing a trait that extends Deserialize and provides a method that returns the delimiter (or sth similar), but that would introduce even more boilerplate.