Have Thunk leverage generic associated types
To be able to have a Thunk own unprocessed data, we need to be able to
have the ThunkContent::from_unprocessed (now
ThunkProcessor::from_unprocessed) be able to process arbitrarily
borrowed input. This means that the signature of the funciton has to be
fn<'a>(&'a str) -> Output<'a>, where Output<'a> is a GAT.

There are some troubles with GATs causing lifetimes to be invariant,
which is why we now pull in VariantPartialEq as a drop-in replacement
for PartialEq derives.