Add #[derive(Dash)] proc-macro
Adds a minimal version of a derive macro that can generate dash-rs'
"internal" representations of GD structs. For now, it only works for
structs where every field in the public API is a 1:1 mapping of some
index in the Geometry Dash format.

The proc-macro implements the `Dash` trait, which will replace the
`HasRobtopFormat` trait. This trait allows (de)serialization based on
an indexed data format using a generic (de)serializer. This decouples
the derive from the string-based format used in GD's HTTP responses,
meaning the Dash macro can also be used for types that describe indexed
objects from, say, the plist save files. The old string-based API moved
into a new subtrait `GJFormat`, which can be implemented for types
implementing `Dash`, and which provides the old APIs from
`HasRobtopFormat`.